+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Art update for Hall of Records
 - Adjusted appends to remove load order dependency with Graham's Things and Anvil of Destiny
+- Adjusted some fixed pixel scenes for Apotheosis map
 
 
 ## 0.6.0